pyproject.toml
[tool.poetry]
name = "chatgpt-terminal"
version = "0.1.0"
description = ""
authors = ["Pedro Salomão <pedro-svm@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "0.28.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
